+++ duat-core/src/text/reader.rs
@@ -1,6 +1,6 @@
 use std::ops::Range;
 
-use super::{Text};
+use super::Text;
 use crate::text::Change;
 
 /// A [`Text`] reader, modifying it whenever a [`Change`] happens
@@ -9,7 +9,7 @@ pub trait Reader: Send + Sync + 'static {
     fn before_change(&mut self, text: &Text, change: Change<&str>) {}
 
     /// Returns ranges that must be updated after a [`Change`]
-    fn after_change(&mut self, text: &Text, change: Change<&str>) -> Vec<Range<u32>>;
+    fn after_change(&mut self, text: &Text, change: Change<&str>) -> Vec<Range<usize>>;
 
     /// Updates a given [`Range`]
     ///
@@ -18,5 +18,5 @@ pub trait Reader: Send + Sync + 'static {
     ///
     /// Must return a [`Vec`] of ranges that have been checked, so
     /// those can be removed from the checking list.
-    fn update_range(&mut self, text: &mut Text, within: Range<u32>);
+    fn update_range(&mut self, text: &mut Text, within: Range<usize>);
 }
